+++ lib/ace/governor/supervisor.ex
@@ -1,3 +1,4 @@
+# TODO remove in breaking update
 defmodule Ace.Governor.Supervisor do
   @moduledoc """
   Supervise governors to maintain a pool of accepting servers.