+++ test/ace/application_test.exs
@@ -1,5 +1,0 @@
-defmodule Ace.ApplicationTest do
-  use ExUnit.Case
-  import Ace.Application
-  doctest Ace.Application
-end